+++ XCL-Charts-demo/src/com/demo/xclcharts/view/BarChart04View.java
@@ -192,18 +192,18 @@ public class BarChart04View extends DemoView {
 			
 			if(v <= 18.5d ) //适中
 			{
-				dataColorA.add((int)Color.rgb(77, 184, 73));				
+				dataColorA.add(Color.rgb(77, 184, 73));				
 			}else if(v <= 24d){ //超重
-				dataColorA.add((int)Color.rgb(252, 210, 9));
+				dataColorA.add(Color.rgb(252, 210, 9));
 			}else if(v <= 27.9d){ //偏胖
-				dataColorA.add((int)Color.rgb(171, 42, 96));	
+				dataColorA.add(Color.rgb(171, 42, 96));	
 			}else{  //肥胖
-				dataColorA.add((int)Color.RED);
+				dataColorA.add(Color.RED);
 			}
 		}  
 		//此地的颜色为Key值颜色及柱形的默认颜色
 		BarData BarDataA = new BarData("",dataSeriesA,dataColorA,
-										(int)Color.rgb(53, 169, 239));
+										Color.rgb(53, 169, 239));
 		
 		chartData.add(BarDataA);
 	}
@@ -226,10 +226,10 @@ public class BarChart04View extends DemoView {
 	 */
 	private void chartDesireLines()
 	{							
-		mCustomLineDataset.add(new CustomLineData("适中",18.5d,(int)Color.rgb(77, 184, 73),3));		
-		mCustomLineDataset.add(new CustomLineData("超重",24d,(int)Color.rgb(252, 210, 9),4));
-		mCustomLineDataset.add(new CustomLineData("偏胖",27.9d,(int)Color.rgb(171, 42, 96),5));	
-		mCustomLineDataset.add(new CustomLineData("肥胖",30d,(int)Color.RED,6));
+		mCustomLineDataset.add(new CustomLineData("适中",18.5d,Color.rgb(77, 184, 73),3));		
+		mCustomLineDataset.add(new CustomLineData("超重",24d,Color.rgb(252, 210, 9),4));
+		mCustomLineDataset.add(new CustomLineData("偏胖",27.9d,Color.rgb(171, 42, 96),5));	
+		mCustomLineDataset.add(new CustomLineData("肥胖",30d,Color.RED,6));
 								
 	}
